refactor: move locales service to common api
