+++ server/zanata-war/src/main/java/org/zanata/security/SecurityFunctions.java
@@ -34,11 +34,10 @@ import org.zanata.model.HProject;
 import org.zanata.model.HProjectIteration;
 import org.zanata.model.LocaleRole;
 import org.zanata.model.ProjectRole;
-import org.zanata.rest.editor.service.resource.LocalesResource;
+import org.zanata.rest.service.LocalesResource;
 import org.zanata.security.annotations.Authenticated;
 import org.zanata.security.permission.GrantsPermission;
 import org.zanata.security.permission.PermissionProvider;
-import org.zanata.util.HttpUtil;
 import org.zanata.util.ServiceLocator;
 import javax.annotation.Nonnull;
 import javax.inject.Inject;
