+++ server/zanata-war/src/main/java/org/zanata/service/impl/LocaleServiceImpl.java
@@ -34,6 +34,7 @@ import javax.enterprise.context.RequestScoped;
 import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 import javax.inject.Named;
+
 import org.apache.deltaspike.jpa.api.transaction.Transactional;
 import org.zanata.common.LocaleId;
 import org.zanata.dao.LocaleDAO;
@@ -430,15 +431,8 @@ public class LocaleServiceImpl implements LocaleService {
                 localeId);
     }
 
-    public static LocaleDetails convertToDTO(HLocale hLocale, String alias) {
-        return new LocaleDetails(hLocale.getLocaleId(),
-                hLocale.retrieveDisplayName(), alias,
-                hLocale.retrieveNativeName(), hLocale.isActive(),
-                hLocale.isEnabledByDefault(), hLocale.getPluralForms());
-    }
-
     public static LocaleDetails convertToDTO(HLocale hLocale) {
-        return convertToDTO(hLocale, null);
+        return LocaleService.convertToDTO(hLocale, null);
     }
 
     public static HLocale convertToHLocale(LocaleDetails localeDetails) {
